+++ src/game_object.cpp
@@ -1,0 +1,14 @@
+class Game_Object {
+
+
+public:
+
+
+protected:
+    //template <class T>
+    //notify_listeners(T* listeners, 
+
+
+private:
+
+};
