+++ src/main.cpp
@@ -1,54 +1,12 @@
 #include <stdio.h>
 #include <string>
-#include <map>
-#include <iostream>
-#include <vector>
-
-
-class Missile {
-public:
-    Missile(std::string missile_name) {
-        name = missile_name;
-    }
-    
-    std::string print_name() {
-        return name;
-    }
-
-private:
-    std::string name;
-};
-
+#include "missile_launcher.h"
+#include "observable.h"
+#include "missile.h"
+#include "missile_launcher_listener.h"
 
 
 //Interface
-class MissileLauncherListener {
-
-public: 
-    virtual void notify_missile_launched(Missile *missile) = 0;
-
-};
-
-class MissileLauncher {
-
-public:
-    Missile * create_missile(std::string name) {
-        Missile *return_missile = new Missile(name);
-        for (MissileLauncherListener * listener : listeners) { 
-            listener->notify_missile_launched(return_missile);
-        }
-
-        return  return_missile;
-    }
-
-    void add_listener(MissileLauncherListener *listener) {
-        listeners.push_back(listener);
-    }
-
-
-private:
-    std::vector<MissileLauncherListener *> listeners;
-};
 
 
 class NosyPerson : public MissileLauncherListener {
@@ -78,7 +36,6 @@ int main () {
     NosyPerson * enoch =  new NosyPerson("enoch has a big nose");
     my_missile_launcher->add_listener(enoch);  
     Missile *big_missile = my_missile_launcher->create_missile("BIG MISSILE!");
-
 
     return 0;
 }
