+++ src/missile_launcher_listener.h
@@ -1,0 +1,11 @@
+#pragma once
+#include "observable.h"
+#include "missile.h"
+#include <string> 
+
+class MissileLauncherListener {
+
+public: 
+    virtual void notify_missile_launched(Missile *missile) = 0;
+
+};
