+++ src/observable.h
@@ -1,0 +1,16 @@
+#pragma once 
+#include <vector> 
+
+
+template <class T>
+class Observable {
+
+public:
+    void add_listener(T *listener) {
+        listeners.push_back(listener);
+    }
+
+protected:
+    std::vector<T *> listeners;
+
+};
